+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - **BREAKING** fee module now designed as constructor, which accepts current network fee values and returns instance with method `getFeeValue`
 - add `payloadLength` param to `getFeeValue`
 - **BREAKING** TX_TYPE.EDIT_COIN_OWNER renamed to TX_TYPE.EDIT_TICKER_OWNER
+- **BREAKING** remove deprecated `numToBig` and default export from converter module, use `numberToBig` and named exports instead
 
 ## 0.20.0 - 2020.12.30
 - **BREAKING** change fee calculation for Create Coin tx (now it depends on unit value)